eliminated test folder and moved code in. next will remove TestApp
