+++ src/notes/FluxNotesEditor.jsx
@@ -2,51 +2,26 @@ import React from 'react';
 import Slate from 'slate';
 import Lang from 'lodash';
 import ContextPortal from '../context/ContextPortal';
-//import ContextManager from '../context/ContextManager';
-
 // versions 0.20.3-0.20.7 of Slate seem to have an issue.
 // when we change the selection and give focus in our key handlers, Slate changes the selection including
 // focus and then immediately takes focus away. Not an issue in 0.20.2 and older. package.json currently
 // forces a version less than 0.20.3.
+// This issue should no longer affect us in our current approach. consider allowing newer version of Slate
 import {Row, Col} from 'react-flexbox-grid';
 import EditorToolbar from './EditorToolbar';
-// Material UI component imports
 import Paper from 'material-ui/Paper';
 import Divider from 'material-ui/Divider';
-
 import AutoReplace from 'slate-auto-replace'
-import SuggestionsPlugin from './../lib/slate-suggestions-dist'
-
+import SuggestionsPlugin from '../lib/slate-suggestions-dist'
 import StructuredFieldPlugin from './StructuredFieldPlugin';
-
-// Styling
 import './FluxNotesEditor.css';
 
 // This forces the initial block to be inline instead of a paragraph. When insert structured field, prevents adding new lines
-const initialState = Slate.Raw.deserialize(
-    {
-        nodes: [
-            {
-                kind: 'block',
-                type: 'inline',
-                nodes: [
-                    {
-                        kind: 'text',
-                        ranges: [
-                            {
-                                'text': ""
-                            }
-                        ]
-                    }
-                ]
-            }
-        ]
-    }, {terse: true});
-
+const initialState = Slate.Plain.deserialize('');
 const schema = {
     nodes: {
         paragraph: props => <p {...props.attributes}>{props.children}</p>,
-        heading: props => <h1 {...props.attributes}>{props.children}</h1>
+        heading: props => <h1 {...props.attributes}>{props.children}</h1>,
     },
     marks: {
         bold: (props) => <strong>{props.children}</strong>
@@ -57,14 +32,6 @@ const structuredFieldTypes = [
     {
         name: 'typeStructuredField',
         value: 'structured_field'
-    },
-    {
-        name: 'typeSubfieldDropdown',
-        value: 'sf_subfield_dropdown'
-    },
-    {
-        name: 'typeSubfieldStaticText',
-        value: 'sf_subfield_statictext'
     }
 ]
 
@@ -76,6 +43,7 @@ class FluxNotesEditor extends React.Component {
 		
 		this.didFocusChange = false;
 		
+		this.selectingForShortcut = null;
 		this.onPortalSelection = this.onPortalSelection.bind(this);
 		this.onChange = this.onChange.bind(this);
 		
@@ -90,9 +58,7 @@ class FluxNotesEditor extends React.Component {
 
         // setup structured field plugin
         const structuredFieldPluginOptions = {
-            updateEditorState: this.onEditorUpdate,
-            updateCurrentShortcutValues: this.onCurrentShortcutValuesUpdate,
-            changeCurrentShortcut: props.changeCurrentShortcut
+            contextManager: this.contextManager
         };
         structuredFieldTypes.forEach((type) => {
             const typeName = type.name;
@@ -103,102 +69,130 @@ class FluxNotesEditor extends React.Component {
         this.structuredFieldPlugin = StructuredFieldPlugin(structuredFieldPluginOptions);
 
         // setup suggestions plugin (autocomplete)
-        let suggestionsShortcuts = [];
-        props.shortcutList.forEach((shortcutKey) => {
-            suggestionsShortcuts.push({
-                "key": shortcutKey,
-                "value": "#" + shortcutKey + "[",
-                "suggestion": shortcutKey
-            });
-        });
-        
-        this.suggestionsPluginShortcuts = SuggestionsPlugin({
+        this.suggestionsPluginCreators = SuggestionsPlugin({
             trigger: '#',
             capture: /#([\w]*)/,
-            suggestions: suggestionsShortcuts,
-            onEnter: (suggestion) => {
-                //console.log("in onEnter");
-                const { state } = this.state; 
-                const { anchorText, anchorOffset } = state
-                const text = anchorText.text
-
-                let index = { start: anchorOffset - 1, end: anchorOffset }
-
-                if (text[anchorOffset - 1] !== '#') {
-                    index = getIndexRangeForCurrentWord(text, anchorOffset - 1, anchorOffset - 1, '#')
-                }
-                    
-                const newText = `${text.substring(0, index.start)} `
-                let transformBeforeInsert = state.transform().deleteBackward(anchorOffset).insertText(newText);
-                let transformAfterInsert = this.insertStructuredFieldTransform(transformBeforeInsert, suggestion.key).focus();
-                return transformAfterInsert.apply();
-            }
+            suggestions: this.suggestionFunction.bind(this),
+            onEnter: this.choseSuggestedShortcut.bind(this)
         });
-
-        // setup suggestions plugin (autocomplete)
-        let suggestionsInsertions = [];
-        //let insertersMap = {};
-        props.inserters.forEach((inserter) => {
-            //insertersMap[inserter.trigger] = inserter.value;
-            suggestionsInsertions.push({
-                "key": inserter.trigger,
-                "value": "@" + inserter.trigger,
-                "suggestion": inserter.trigger,
-                "valueFunc" : inserter.value
-            });
-        });
-        
-        this.suggestionsPluginInsertions = SuggestionsPlugin({
+        this.suggestionsPluginInserters = SuggestionsPlugin({
             trigger: '@',
             capture: /@([\w]*)/,
-            suggestions: suggestionsInsertions,
-            onEnter: (suggestion) => {
-                //console.log("in onEnter " + suggestion.key);
-				let transform = this.handleInserter(true, suggestion.valueFunc);
-				if (Lang.isNull(transform)) {
-					return this.state.state;
-				} else {
-					return transform.apply();
-				}
-            }
+            suggestions: this.suggestionFunction.bind(this),
+            onEnter: this.choseSuggestedShortcut.bind(this)
         });
         
-       // do not use onKeyDown, use auto-replace plugin, add to existing global 'plugins' list
         this.plugins = [
-			this.suggestionsPluginShortcuts,
-			this.suggestionsPluginInsertions,
+			this.suggestionsPluginCreators,
+			this.suggestionsPluginInserters,
 			this.structuredFieldPlugin
 		];
-				
-		// now add an AutoReplace plugin instance for each shortcut we're supporting as well
-		props.shortcutList.forEach((shortcutKey) => {
-			this.plugins.push(AutoReplace({
-				"trigger": "[",
-				"before": new RegExp("(#" + shortcutKey + ")", "i"),
-				"transform": (transform, e, data, matches) => {
-					// need to use Transform object provided to this method, which AutoReplace .apply()s after return.
-					return this.insertStructuredFieldTransform(transform, shortcutKey);
-				}
-			}));
-		});
 		
-		// now add an AutoReplace plugin instance for each inserter we're supporting
-		props.inserters.forEach((inserter) => {
-			this.plugins.push(AutoReplace({
-				"trigger": 'space',
-				"before": new RegExp("(@" + inserter.trigger + ")", "i"),
-				"transform": (transform, e, data, matches) => {
-					let newTransform = this.handleInserter(false, inserter.value, transform);
-					if (Lang.isNull(newTransform)) {
-						newTransform = transform;
-					}
-					return newTransform;
-				}
-			}));
+		this.autoReplaceBeforeRegExp = undefined;
+		
+		let autoReplaceAfters = [];
+		let allShortcuts = this.props.shortcutManager.getAllShortcutClasses();
+		allShortcuts.forEach((shortcutC) => {
+			autoReplaceAfters = autoReplaceAfters.concat(shortcutC.getTriggers());			
 		});
+		this.autoReplaceBeforeRegExp = new RegExp("(" + autoReplaceAfters.join("|") + ")", 'i');
+		
+		// now add an AutoReplace plugin instance for each shortcut we're supporting as well
+		this.plugins.push(AutoReplace({
+			"trigger": 'space',
+			"before": this.autoReplaceBeforeRegExp,
+			"transform": this.autoReplaceTransform.bind(this)
+		}));
     }
+		
+	suggestionFunction(text) {
+		let shortcuts = this.contextManager.getCurrentlyValidShortcuts();
+        let suggestionsShortcuts = [];
+        shortcuts.forEach((shortcut) => {
+			const triggers = shortcut.getTriggers();
+			triggers.forEach((trigger) => {
+				const triggerNoPrefix = trigger.substring(1);
+				if (triggerNoPrefix.includes(text)) {
+					suggestionsShortcuts.push({
+						"key": triggerNoPrefix,
+						"value": trigger,
+						"suggestion": triggerNoPrefix
+					});
+				}
+			});
+        });
+		return suggestionsShortcuts;
+	}
 	
-	suggestionDeleteExistingTransform(transform = null) {
+	choseSuggestedShortcut(suggestion) {
+		const { state } = this.state; 
+        let shortcut = this.props.newCurrentShortcut(suggestion.value);
+		
+		if (!Lang.isNull(shortcut) && shortcut.needToSelectValueFromMultipleOptions()) {
+			return this.openPortalToSelectValueForShortcut(shortcut, true, state.transform()).apply();
+		} else {
+            let transformBeforeInsert = this.suggestionDeleteExistingTransform(state.transform(), shortcut.getPrefixCharacter());
+			let transformAfterInsert = this.insertStructuredFieldTransform(transformBeforeInsert, shortcut).focus();
+			return transformAfterInsert.apply();
+		}
+	}
+	
+	insertShortcut(shortcutTrigger, transform = undefined) {
+		if (Lang.isUndefined(transform)) {
+			transform = this.state.state.transform();
+		}
+		let shortcut = this.props.newCurrentShortcut(shortcutTrigger);
+		if (!Lang.isNull(shortcut) && shortcut.needToSelectValueFromMultipleOptions()) {
+			return this.openPortalToSelectValueForShortcut(shortcut, false, transform);
+		} else {
+			return this.insertStructuredFieldTransform(transform, shortcut).collapseToStartOfNextText().focus();
+		}
+	}
+	
+	autoReplaceTransform(transform, e, data, matches) {
+		// need to use Transform object provided to this method, which AutoReplace .apply()s after return.
+		this.insertShortcut(matches.before[0], transform);
+	}
+	
+	openPortalToSelectValueForShortcut(shortcut, needToDelete, transform) {
+		let portalOptions = shortcut.getValueSelectionOptions();
+		let pos = position(this.state);
+		
+		this.setState({
+			isPortalOpen: true,
+			portalOptions: portalOptions,
+			needToDelete: needToDelete,
+			left: pos.left,
+			top: pos.top
+		});
+		this.selectingForShortcut = shortcut;
+		return transform.blur();
+	}
+
+	// called from portal when an item is selected (context is not null) or if portal is closed without
+	// selection (context is null)
+	onPortalSelection = (state, context) => {
+		this.setState({ isPortalOpen: false });
+		if (Lang.isNull(context)) return state;
+		let shortcut = this.selectingForShortcut;
+		this.selectingForShortcut = null;
+		shortcut.clearValueSelectionOptions();
+		shortcut.setText(context.context);
+		if (shortcut.isContext()) {
+			shortcut.setValueObject(context.object);
+		}
+		this.contextManager.contextUpdated();
+		let transform;
+		if (this.state.needToDelete) {
+			transform = this.suggestionDeleteExistingTransform(null, shortcut.getPrefixCharacter());
+		} else {
+			transform = this.state.state.transform();
+		}
+		return this.insertStructuredFieldTransform(transform, shortcut).collapseToStartOfNextText().focus().apply();
+	}
+
+	// consider reusing this method to replace code in choseSuggestedShortcut function
+	suggestionDeleteExistingTransform(transform = null, prefixCharacter) {
 		const { state } = this.state;
 		if (Lang.isNull(transform)) {
 			transform = state.transform();
@@ -208,8 +202,8 @@ class FluxNotesEditor extends React.Component {
 
 		let index = { start: anchorOffset - 1, end: anchorOffset }
 
-		if (text[anchorOffset - 1] !== '@') {
-			index = getIndexRangeForCurrentWord(text, anchorOffset - 1, anchorOffset - 1, '@')
+		if (text[anchorOffset - 1] !== prefixCharacter) {
+			index = getIndexRangeForCurrentWord(text, anchorOffset - 1, anchorOffset - 1, prefixCharacter)
 		}
 			
 		const newText = `${text.substring(0, index.start)}`
@@ -217,119 +211,24 @@ class FluxNotesEditor extends React.Component {
 			.deleteBackward(anchorOffset)
 			.insertText(newText)	
 	}
-
-	// debugging @condition which is case where valueFunc returns an array so we present a pick list via 
-	// ContextPortal
-	// autocomplete call to handleInserter:
-	//		let transform = this.handleInserter(true, suggestion.valueFunc);
-	//		works if click on condition in autocomplete then hit enter to choose 1st condition
-	//		works if hit enter to choose condition autocomplete then hit enter to choose 1st condition
-	// autoreplace call to handleInserter:
-	//		let newTransform = this.handleInserter(false, inserter.value, transform);
-	//		get extra enter in editor before text if:
-	//			type @condition followed by a space then hit enter to choose 1st condition
-	//   		solved by returning blur transform from array case
-	//		now enter/enter case doesn't work (autcomplete with enter then enter to choose condition):
-	//			doesn't get inserted but may because of crash in suggestions plugin
-	//			remove setting menu to null on close of portal
-	//		now losing state typed for autocomplete case
-	//			looks like it happens if I just hit enter enter. Those enter keys are consumed as well as the @cond...
-	//			maybe the get word functionality is getting confused when there is no space character to find? have an
-	//			@ though which should be enough
-	//			also, need to get rid of TypeErros from slate suggestions (which Dylan has done in our own version)
 	
-	handleInserter(needToDelete, valueFunc, transform = null) {
-		let value = valueFunc(this.contextManager); //this.props.patient
-		if (Lang.isArray(value)) {
-			let portalOptions = [];
-			value.forEach((item) => {
-				portalOptions.push({key: item.entryId, context: item.specificType.coding.displayText, object: item});
-			});
-			let pos = position(this.state);
-			
-			let newTransform = transform;
-			if (Lang.isNull(transform)) {
-				newTransform = this.state.state.transform();
-			}
-			this.setState({
-				isPortalOpen: true,
-				portalOptions: portalOptions,
-				needToDelete: needToDelete,
-				left: pos.left,
-				top: pos.top
-			});
-			return newTransform.blur();
-		} else {
-			value = "" + value;
-			if (needToDelete) {
-				return this.suggestionDeleteExistingTransform(transform).insertText(value).focus();
-			} else {
-				let newTransform = transform;
-				if (Lang.isNull(transform)) {
-					newTransform = this.state.state.transform();
-				}
-				return newTransform.insertText(value).focus();
-			}
-		}
-	}
-	
-    insertStructuredFieldTransform(transform, shortcutType) {
-        let shortcut = this.props.newCurrentShortcut(shortcutType);
+    insertStructuredFieldTransform(transform, shortcut) {
 		if (Lang.isNull(shortcut)) return transform.focus();
         let result = this.structuredFieldPlugin.transforms.insertStructuredField(transform, shortcut); //2nd param needs to be Shortcut Object, how to create?
-        let transformAfterInsert = result[0];
-        return transformAfterInsert;
-    }
-
-    onEditorUpdate = (newState) => {
-		let curSelection = this.state.state.selection;
-        //console.log(`Plugin updating state`);
-        this.setState({
-            state: newState
-        });
-		//console.log(newState);
-		if (!Lang.isEqual(curSelection, newState.selection)) this.onSelectionChange(newState.selection, newState);
-    }
-    onCurrentShortcutValuesUpdate = (name, value) => {
-        this.props.currentShortcut.updateValue(name, value, false);
-    }
-
-	componentDidMount = () => {
-	}
-	
-    componentDidUpdate = (prevProps, prevState) => {
+        return result[0];
     }
 
     onChange = (state) => {
-        //console.log("onChange: START");
         this.setState({
             state: state
         });
-        //console.log("onChange: DONE");
     }
-
-    onSelectionChange = (selection, state) => {
-        this.didFocusChange = this.structuredFieldPlugin.transforms.onSelectionChange(selection, state);
-		//console.log("onSelectionChange. did focus change = " + this.didFocusChange);
-    }
-	
-	onBlur = (event, data, state, editor) => {
-		//console.log("onBlur: state.selection.startKey=" + state.selection.startKey);
-		if (this.didFocusChange) {
-			this.didFocusChange = false;
-			event.preventDefault();
-			//console.log("onBlur: suppress blur. DONE");
-			return state;
-		}
-		//console.log("onBlur: let core handle blur. DONE");
-		return;
-	}
 
     // This gets called when the before the component receives new properties
     componentWillReceiveProps = (nextProps) => {
-
         if (this.props.itemToBeInserted !== nextProps.itemToBeInserted) {
             this.handleSummaryUpdate(nextProps.itemToBeInserted);
+            this.props.itemInserted();
         }
     }
 
@@ -337,14 +236,28 @@ class FluxNotesEditor extends React.Component {
      * Handle updates when we have a new
      */
     handleSummaryUpdate = (itemToBeInserted) => {
-        const currentState = this.state.state;
-        const state = currentState
-            .transform()
-            .insertText(itemToBeInserted)
-            .focus()
-            .apply();
-        this.setState({state: state});
-    }
+		let state;
+		const currentState = this.state.state;
+		let transform = currentState.transform();
+		let regExp = new RegExp("(@[\\w\\-]+|#[\\w\\-]+)", "i");
+		let result = regExp.exec(itemToBeInserted);
+		let remainder = itemToBeInserted;
+		let start, before;
+		while (!Lang.isNull(result)) {
+			start = remainder.indexOf(result[0]);
+			before = remainder.substring(0, start);
+			remainder = remainder.substring(start + result[0].length);
+			transform = transform
+				.insertText(before)
+			transform = this.insertShortcut(result[0], transform);
+			result = regExp.exec(remainder);
+		}
+		if (remainder.length > 0) {
+			transform = transform.insertText(remainder);
+		}
+		state = transform.focus().apply();
+		this.setState({state: state});
+	}
 
     /**
      * Check if the current selection has a mark with `type` in it.
@@ -362,33 +275,10 @@ class FluxNotesEditor extends React.Component {
         return state.blocks.some(node => node.type === type);
     }
 	
-	// called from portal when an item is selected (context is not null) or if portal is closed without
-	// selection (context is null)
-	onPortalSelection = (state, context) => {
-		//console.log("onPortalSelection for context portal. context is null: " + (Lang.isNull(context)));
-		//try {
-			this.setState({ isPortalOpen: false });
-		//} catch (e) {
-		//	console.log("TypeError consumed");
-		//}
-		if (Lang.isNull(context)) return state;
-		//console.log("onPortalSelection for context portal " + context.context + " needToDelete some text first=" + this.state.needToDelete);
-		this.contextManager.addContext(context.object);
-		let transform;
-		if (this.state.needToDelete) {
-			transform = this.suggestionDeleteExistingTransform();
-		} else {
-			transform = this.state.state.transform();
-		}
-		return transform.insertText(context.context).focus().apply();
-	}
-
     render = () => {
-        //let {state} = this.state;
-        //let isStructuredField = structuredFieldPlugin.utils.isSelectionInStructuredField(state);
-        // Extract portal component from the plugin 
-        const ShortcutsPortal = this.suggestionsPluginShortcuts.SuggestionPortal;
-        const InsertionsPortal = this.suggestionsPluginInsertions.SuggestionPortal;
+        const CreatorsPortal = this.suggestionsPluginCreators.SuggestionPortal;
+		const InsertersPortal = this.suggestionsPluginInserters.SuggestionPortal;
+
         let noteDescriptionContent = null;
         if (this.props.patient == null) {
             noteDescriptionContent = "";
@@ -449,14 +339,12 @@ class FluxNotesEditor extends React.Component {
                             plugins={this.plugins}
                             state={this.state.state}
                             onChange={this.onChange}
-                            onSelectionChange={this.onSelectionChange}
-                            onBlur={this.onBlur}
                             schema={schema}
                         />
 						{errorDisplay}
-						<ShortcutsPortal 
+						<CreatorsPortal 
 							state={this.state.state} />
-						<InsertionsPortal
+						<InsertersPortal
 							state={this.state.state} />
 						<ContextPortal
 							left={this.state.left}
@@ -474,12 +362,10 @@ class FluxNotesEditor extends React.Component {
                 </Paper>
             </div>
         );
-
     }
 }
 
 function getPos(el) {
-    // yay readability
     for (var lx=0, ly=0;
          el != null;
          lx += el.offsetLeft, ly += el.offsetTop, el = el.offsetParent);
@@ -497,20 +383,19 @@ function position(state) {
 }
 
 function getIndexRangeForCurrentWord(text, index, initialIndex, initialChar) {
-  if (index === initialIndex) {
-    return { 	start: getIndexRangeForCurrentWord(text, index - 1, initialIndex, initialChar), 
-				end: getIndexRangeForCurrentWord(text, index + 1, initialIndex, initialChar) 	}
-  }
-  //console.log("start?" + (index<initialIndex) + " char=" + text[index] + " initial char=" + initialChar);
-  if (text[index] === " " || text[index] === initialChar || text[index] === undefined) {
-    return index
-  }
-  if (index < initialIndex) {
-    return getIndexRangeForCurrentWord(text, index - 1, initialIndex, initialChar)
-  }
-  if (index > initialIndex) {
-    return getIndexRangeForCurrentWord(text, index + 1, initialIndex, initialChar)
-  }
+    if (index === initialIndex) {
+        return { 	start: getIndexRangeForCurrentWord(text, index - 1, initialIndex, initialChar), 
+                    end: getIndexRangeForCurrentWord(text, index + 1, initialIndex, initialChar) 	}
+    }
+    if (text[index] === " " || text[index] === initialChar || text[index] === undefined) {
+        return index
+    }
+    if (index < initialIndex) {
+        return getIndexRangeForCurrentWord(text, index - 1, initialIndex, initialChar)
+    }
+    if (index > initialIndex) {
+        return getIndexRangeForCurrentWord(text, index + 1, initialIndex, initialChar)
+    }
 }
 
 export default FluxNotesEditor;
