+++ src/shortcuts/NameInserter.jsx
