+++ src/shortcuts/PatientInserter.jsx
