+++ src/shortcuts/ProgressionCreator.jsx
@@ -1,13 +1,13 @@
 // React Imports:
 import React from 'react';
 import CreatorShortcut from './CreatorShortcut';
-import ProgressionForm from '../../forms/ProgressionForm';
+import ProgressionForm from '../forms/ProgressionForm';
 import ProgressionStatusCreator from './ProgressionStatusCreator';
 import ProgressionReasonsCreator from './ProgressionReasonsCreator';
-import Patient from '../../patient/Patient';
+import Patient from '../patient/Patient';
 import Lang from 'lodash'
 import moment from 'moment';
-import progressionLookup from '../../lib/progression_lookup';
+import progressionLookup from '../lib/progression_lookup';
 
 class ProgressionCreator extends CreatorShortcut {
     // onUpdate is passed from React components that need to be notified when the progression value(s) change
