+++ src/shortcuts/StageInserter.jsx
