+++ src/shortcuts/StagingMCreator.jsx
@@ -1,8 +1,8 @@
 import CreatorShortcut from './CreatorShortcut';
 
-const lookup = require('../../lib/staging_lookup');
+const lookup = require('../lib/staging_lookup');
 
-export default class StagingNCreator extends CreatorShortcut {
+export default class StagingMCreator extends CreatorShortcut {
     constructor(onUpdate, obj) {
         super();
 	}
@@ -11,14 +11,14 @@ export default class StagingNCreator extends CreatorShortcut {
 		super.initialize(contextManager, trigger);
 		this.text = trigger;
 		this.parentContext = contextManager.getActiveContextOfType("#staging");
-		this.parentContext.setAttributeValue("N", trigger.substring(1), false);
+		this.parentContext.setAttributeValue("M", trigger.substring(1), false);
         this.parentContext.addChild(this);
 	}
-
+	
     onBeforeDeleted() {
         let result = super.onBeforeDeleted();
         if (result) {
-            this.parentContext.setAttributeValue("N", "", false);
+            this.parentContext.setAttributeValue("M", "", false);
             this.parentContext.removeChild(this);
         }
         return result;
@@ -29,25 +29,26 @@ export default class StagingNCreator extends CreatorShortcut {
 	}
 	
     getShortcutType() { 
-        return "#staging-n";
+        return "#staging-m";
     }
 
 	validateInCurrentContext(contextManager) {
 		let errors = [];
 		if (!contextManager.isContextOfTypeActive("#staging")) {
-			errors.push("Staging N values invalid without #staging. Use #staging to add a new staging to your narrative.");
+			errors.push("Staging M values invalid without #staging. Use #staging to add a new staging to your narrative.");
             return errors;
 		}
 		let parentContext = contextManager.getActiveContextOfType("#staging");
-		if (parentContext.getAttributeValue("N").length > 0) {
-			errors.push("Staging N value already specified. Only one N value allowed per staging.");
+		if (parentContext.getAttributeValue("M").length > 0) {
+			errors.push("Staging M value already specified. Only one M value allowed per staging.");
 		}
 		return errors;
 	}
+	
 	static getTriggers() {
-		const ns = lookup.getNsNamesForEdition(7);
+		const ms = lookup.getMsNamesForEdition(7);
 		let result = [];
-		ns.forEach((val) => {
+		ms.forEach((val) => {
 			result.push("#" + val);
 		});
 		return result;
