+++ src/shortcuts/StagingTCreator.jsx
@@ -1,5 +1,5 @@
 import CreatorShortcut from './CreatorShortcut';
-const lookup = require('../../lib/staging_lookup');
+const lookup = require('../lib/staging_lookup');
 
 export default class StagingTCreator extends CreatorShortcut {
     constructor(onUpdate, obj) {
