--- src/test/context/ContextPortal.css
@@ -1,25 +1,0 @@
-.context-portal {
-  display: none;
-  position: absolute;
-  z-index: 10000;
-  margin-left: 20px;
-  margin-top: 30px;
-  width: 200px;
-}
-
-.context-portal ul {
-  list-style-type: none;
-  margin-top: 0;
-  position: relative;
-  padding-left: 0;
-  margin-left: 5px;
-  background-color: #fff;
-}
-
-.context-portal li {
-  padding: 5px 10px;
-}
-
-.context-portal li.selected {
-  background-color: #eee;
-}
